+++ src/app/api/admin/brands/[id]/route.js
@@ -2,11 +2,15 @@ import connectDB from "@/lib/db";
 import Brand from "@/models/Brand";
 import { ok, notFound, serverError } from "@/lib/apiResponse";
 import slugify from "slugify";
+import { deleteModelImage, updateImage } from "@/lib/cloudinaryDelete";
 
 export async function PUT(request, { params }) {
   try {
     const { name, status, logo } = await request.json();
     await connectDB();
+    const oldBrand = await updateImage(Brand, params.id, logo);
+    if (!oldBrand) return notFound("Brand not found");
+
     const brand = await Brand.findById(params.id);
     if (!brand) return notFound("Brand not found");
     if (name) {
@@ -25,6 +29,11 @@ export async function PUT(request, { params }) {
 export async function DELETE(_, { params }) {
   try {
     await connectDB();
+
+    const oldBrand = await deleteModelImage(Brand, params.id);
+
+    if (!oldBrand) return notFound("Brand not found");
+
     const brand = await Brand.findByIdAndDelete(params.id);
     if (!brand) return notFound("Brand not found");
     return ok(null, "Brand deleted");
